Bugfix: force floats
We got something!

Edit: signed, 8-bit, PCM, big endian, mono, 44.1kHz data, by the
looks of it!  \o/

+++ columnstosines.c
@@ -7,7 +7,7 @@ a series of sine waves.  */
 #include <stdint.h>
 #include <stdio.h>
 
-int i;
+float i;
 int freq;
 int col;
 int row;
@@ -28,7 +28,7 @@ void convert(FILE *inputFilePointer, FILE *outputFilePointer) {
 				if (col & (1 << (7 - row))) { /* Lowest frequency oscillator first */
 					freq = 16000 + (500 * row); /* Hardwire each pixel height as a single sine wave "beam" 500Hz apart from its neighbours, starting at 16kHz, for now */
 					/* Each sine wave should be 1/9th volume, for mixing with headroom */
-					mix += (1 / 9) * sin(freq * (i / 44100) * M_PI_2); /* The number of cycles per second is multiplied by the number of seconds.  Even though the latter's between 0 and 0.25, the frequencies bring it up.  Hardwire CD quality sample rate for now. */
+					mix += (1.0 / 9.0) * sin(freq * (i / 44100.0) * M_PI_2); /* The number of cycles per second is multiplied by the number of seconds.  Even though the latter's between 0 and 0.25, the frequencies bring it up.  Hardwire CD quality sample rate for now. */
 				}
 			}
